originators: split into smaller files
